Complete the setting module; Complete the add spending module: add a spending, add a spending from a card. Complete the reporting module: all spending, all spending from a card; Monthly Summary. Going to add charts and more summary in report module; and going to add data backup function and data restore module.

+++ schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE cards (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     bank TEXT NOT NULL,
     name TEXT UNIQUE NOT NULL,
-    cur_balance INTEGER,
+    cur_balance REAL,
     pay_date INTEGER
 );
 
@@ -36,7 +36,7 @@ CREATE TABLE degrees (
 CREATE TABLE spending (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     name TEXT NOT NULL,
-    amount INTEGER NOT NULL,
+    amount REAL NOT NULL,
     category INTEGER NOT NULL,
     sub_category INTEGER NOT NULL,
     yr INTEGER NOT NULL,
